requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
pytesseract>=0.3.13
Pillow>=10.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
